PI Task: split PI LogDialog into seperate file for easier testing
* This allows it to be run as stand-alone application.
